+++ Sensors/SMCDellSensors/SMCDellSensors.cpp
@@ -9,16 +9,21 @@
 
 #include "SMCDellSensors.hpp"
 #include "KeyImplementations.hpp"
+#include <Headers/plugin_start.hpp>
+#include <Headers/kern_version.hpp>
+#include <IOKit/pwr_mgt/RootDomain.h>
+
+#include "kern_hooks.hpp"
 
 OSDefineMetaClassAndStructors(SMCDellSensors, IOService)
-
-bool ADDPR(debugEnabled) = false;
-uint32_t ADDPR(debugPrintDelay) = 0;
 
 bool SMCDellSensors::init(OSDictionary *properties) {
 	if (!IOService::init(properties)) {
 		return false;
 	}
+
+	if (!SMIMonitor::getShared())
+		SMIMonitor::createShared();
 
 	SMIMonitor::getShared()->fanMult = 1; //linux proposed to get nominal speed and if it high then change multiplier
 	OSNumber * Multiplier = OSDynamicCast(OSNumber, properties->getObject("FanMultiplier"));
@@ -33,16 +38,14 @@ IOService *SMCDellSensors::probe(IOService *provider, SInt32 *score) {
 		SYSLOG("sdell", "failed to probe the parent");
 		return nullptr;
 	}
-	
+
 	if (!SMIMonitor::getShared()->probe())
 		return nullptr;
-	
-	//WARNING: watch out, key addition is sorted here!
-	
+
 	auto fanCount = min(SMIMonitor::getShared()->fanCount, MaxIndexCount);
 	VirtualSMCAPI::addKey(KeyFNum, vsmcPlugin.data,
 		VirtualSMCAPI::valueWithUint8(fanCount, nullptr, SMC_KEY_ATTRIBUTE_CONST | SMC_KEY_ATTRIBUTE_READ));
-	
+
 	for (size_t i = 0; i < fanCount; i++) {
 		VirtualSMCAPI::addKey(KeyF0Ac(i), vsmcPlugin.data, VirtualSMCAPI::valueWithFp(0, SmcKeyTypeFpe2, new F0Ac(i), SMC_KEY_ATTRIBUTE_WRITE | SMC_KEY_ATTRIBUTE_READ));
 		VirtualSMCAPI::addKey(KeyF0Mn(i), vsmcPlugin.data, VirtualSMCAPI::valueWithFp(0, SmcKeyTypeFpe2, new F0Mn(i), SMC_KEY_ATTRIBUTE_WRITE | SMC_KEY_ATTRIBUTE_READ));
@@ -54,9 +57,10 @@ IOService *SMCDellSensors::probe(IOService *provider, SInt32 *score) {
 		VirtualSMCAPI::valueWithUint16(0, new FS__(), SMC_KEY_ATTRIBUTE_WRITE | SMC_KEY_ATTRIBUTE_READ));
 
 	OSArray* fanNames = OSDynamicCast(OSArray, getProperty("FanNames"));
+	OSArray* stopFanOffsets = OSDynamicCast(OSArray, getProperty("StopFanOffsets"));
 	char fan_name[DiagFunctionStrLen];
 
-	for (size_t i = 0, cpu = 0; i < fanCount; i++) {
+	for (unsigned int i = 0, cpu = 0; i < fanCount; i++) {
 		FanTypeDescStruct	desc;
 		FanInfo::SMMFanType type = SMIMonitor::getShared()->state.fanInfo[i].type;
 		if (type == FanInfo::Unsupported) {
@@ -64,17 +68,22 @@ IOService *SMCDellSensors::probe(IOService *provider, SInt32 *score) {
 			DBGLOG("sdell", "Fan type %d is unknown, auto assign value %d", type, auto_type);
 			type = auto_type;
 		}
-		snprintf(fan_name, DiagFunctionStrLen, "Fan %lu", i);
-		if (fanNames) {
+		snprintf(fan_name, DiagFunctionStrLen, "Fan %u", i);
+		if (fanNames && type < fanNames->getCount()) {
 			OSString* name = OSDynamicCast(OSString, fanNames->getObject(type));
 			if (name)
 				lilu_os_strncpy(fan_name, name->getCStringNoCopy(), DiagFunctionStrLen);
+		}
+		if (stopFanOffsets && i < stopFanOffsets->getCount()) {
+			OSNumber* offset = OSDynamicCast(OSNumber, stopFanOffsets->getObject(i));
+			if (offset)
+				SMIMonitor::getShared()->state.fanInfo[i].stopOffset = offset->unsigned32BitValue();
 		}
 		lilu_os_strncpy(desc.strFunction, fan_name, DiagFunctionStrLen);
 		VirtualSMCAPI::addKey(KeyF0ID(i), vsmcPlugin.data, VirtualSMCAPI::valueWithData(
 			reinterpret_cast<const SMC_DATA *>(&desc), sizeof(desc), SmcKeyTypeFds, nullptr, SMC_KEY_ATTRIBUTE_CONST | SMC_KEY_ATTRIBUTE_READ));
 	}
-	
+
 	auto tempCount = min(SMIMonitor::getShared()->tempCount, MaxIndexCount);
 	for (size_t i = 0; i < tempCount; i++) {
 		TempInfo::SMMTempSensorType type = SMIMonitor::getShared()->state.tempInfo[i].type;
@@ -82,7 +91,7 @@ IOService *SMCDellSensors::probe(IOService *provider, SInt32 *score) {
 			DBGLOG("sdell", "Temp sensor type %d is unknown, auto assign value %d", type, SMIMonitor::getShared()->state.tempInfo[i].index);
 			type = static_cast<TempInfo::SMMTempSensorType>(SMIMonitor::getShared()->state.tempInfo[i].index);
 		}
-		
+
 		switch (type)
 		{
 		case TempInfo::CPU:
@@ -114,9 +123,9 @@ IOService *SMCDellSensors::probe(IOService *provider, SInt32 *score) {
 			break;
 		}
 	}
-	
+
 	qsort(const_cast<VirtualSMCKeyValue *>(vsmcPlugin.data.data()), vsmcPlugin.data.size(), sizeof(VirtualSMCKeyValue), VirtualSMCKeyValue::compare);
-	
+
 	return this;
 }
 
@@ -125,13 +134,45 @@ bool SMCDellSensors::start(IOService *provider) {
 		SYSLOG("sdell", "failed to start the parent");
 		return false;
 	}
+
+	setProperty("VersionInfo", kextVersion);
+
+	notifier = registerSleepWakeInterest(IOSleepHandler, this);
+	if (notifier == NULL) {
+		SYSLOG("sdell", "failed to register sleep/wake interest");
+		return false;
+	}
 	
+	if (!eventTimer) {
+		if (!workLoop)
+			workLoop = IOWorkLoop::workLoop();
+
+		if (workLoop) {
+			eventTimer = IOTimerEventSource::timerEventSource(workLoop,
+			[](OSObject *owner, IOTimerEventSource *) {
+				SMIMonitor::getShared()->handlePowerOn();
+			});
+
+			if (eventTimer) {
+				IOReturn result = workLoop->addEventSource(eventTimer);
+				if (result != kIOReturnSuccess) {
+					SYSLOG("sdell", "SMCDellSensors addEventSource failed");
+					OSSafeReleaseNULL(eventTimer);
+				}
+			}
+			else
+				SYSLOG("sdell", "SMCDellSensors timerEventSource failed");
+		}
+		else
+			SYSLOG("sdell", "SMCDellSensors IOService instance does not have workLoop");
+	}
+
 	SMIMonitor::getShared()->start();
-	
+
 	PMinit();
 	provider->joinPMtree(this);
 	registerPowerDriver(this, powerStates, arrsize(powerStates));
-	
+
 	vsmcNotifier = VirtualSMCAPI::registerHandler(vsmcNotificationHandler, this);
 	return vsmcNotifier != nullptr;
 }
@@ -161,43 +202,38 @@ void SMCDellSensors::stop(IOService *provider) {
 
 IOReturn SMCDellSensors::setPowerState(unsigned long state, IOService *whatDevice){
 	DBGLOG("sdell", "changing power state to %lu", state);
-
-	if (state == PowerStateOff)
-		SMIMonitor::getShared()->handlePowerOff();
-	else if (state == PowerStateOn)
-		SMIMonitor::getShared()->handlePowerOn();
-
 	return kIOPMAckImplied;
 }
 
-EXPORT extern "C" kern_return_t ADDPR(kern_start)(kmod_info_t *, void *) {
-	// Report success but actually do not start and let I/O Kit unload us.
-	// This works better and increases boot speed in some cases.
-	PE_parse_boot_argn("liludelay", &ADDPR(debugPrintDelay), sizeof(ADDPR(debugPrintDelay)));
-	ADDPR(debugEnabled) = checkKernelArgument("-vsmcdbg") || checkKernelArgument("-sdelldbg");
-	SMIMonitor::createShared();
-	return KERN_SUCCESS;
+IOReturn SMCDellSensors::IOSleepHandler(void *target, void *, UInt32 messageType, IOService *provider, void *messageArgument, vm_size_t)
+{
+	sleepWakeNote *swNote = (sleepWakeNote *)messageArgument;
+
+	if (messageType != kIOMessageSystemWillSleep &&
+		messageType != kIOMessageSystemHasPoweredOn &&
+		messageType != kIOMessageSystemWillNotSleep) {
+		return kIOReturnUnsupported;
+	}
+
+	DBGLOG("sdell", "IOSleepHandler message type = 0x%x", messageType);
+
+	swNote->returnValue = 0;
+	acknowledgeSleepWakeNotification(swNote->powerRef);
+
+	auto that = OSDynamicCast(SMCDellSensors, reinterpret_cast<SMCDellSensors*>(target));
+
+	if (messageType == kIOMessageSystemWillSleep) {
+		if (that && that->eventTimer)
+			that->eventTimer->cancelTimeout();
+		SMIMonitor::getShared()->handlePowerOff();
+	} else if (messageType == kIOMessageSystemHasPoweredOn || messageType == kIOMessageSystemWillNotSleep) {
+		auto that = OSDynamicCast(SMCDellSensors, reinterpret_cast<SMCDellSensors*>(target));
+		if (that && that->eventTimer)
+			that->eventTimer->setTimeoutMS(4000);
+		else
+			SMIMonitor::getShared()->handlePowerOn();
+	}
+
+	return kIOReturnSuccess;
 }
 
-EXPORT extern "C" kern_return_t ADDPR(kern_stop)(kmod_info_t *, void *) {
-	// It is not safe to unload VirtualSMC plugins!
-	return KERN_FAILURE;
-}
-
-#ifdef __MAC_10_15
-
-// macOS 10.15 adds Dispatch function to all OSObject instances and basically
-// every header is now incompatible with 10.14 and earlier.
-// Here we add a stub to permit older macOS versions to link.
-// Note, this is done in both kern_util and plugin_start as plugins will not link
-// to Lilu weak exports from vtable.
-
-kern_return_t WEAKFUNC PRIVATE OSObject::Dispatch(const IORPC rpc) {
-	PANIC("util", "OSObject::Dispatch smcbat stub called");
-}
-
-kern_return_t WEAKFUNC PRIVATE OSMetaClassBase::Dispatch(const IORPC rpc) {
-	PANIC("util", "OSMetaClassBase::Dispatch smcbat stub called");
-}
-
-#endif