+++ Sensors/SMCDellSensors/SMIMonitor.cpp
@@ -9,22 +9,33 @@
  */
 
 #include "SMIMonitor.hpp"
+#include "kern_hooks.hpp"
+
 #include <Headers/kern_cpu.hpp>
 
 extern "C" {
-#include <Library/osfmk/i386/pmCPU.h>
+#include <i386/pmCPU.h>
 }
 
 SMIMonitor *SMIMonitor::instance = nullptr;
+atomic_bool SMIMonitor::busy = 0;
 
 OSDefineMetaClassAndStructors(SMIMonitor, OSObject)
 
 int SMIMonitor::i8k_smm(SMMRegisters *regs) {
 	int rc;
 	int eax = regs->eax;  //input value
+
+	while (atomic_load_explicit(&KERNELHOOKS::active_output, memory_order_acquire)) { IOSleep(4); }
+
+	atomic_store_explicit(&busy, true, memory_order_release);
 	
-	IOSimpleLockLock(preemptionLock);
-	
+	if (atomic_load_explicit(&KERNELHOOKS::active_output, memory_order_acquire)) {
+		DBGLOG("sdell", "stop accessing smm, active_outputs = %d", atomic_load_explicit(&KERNELHOOKS::active_output, memory_order_acquire));
+		atomic_store_explicit(&busy, false, memory_order_release);
+		return -1;
+	}
+
 #if __LP64__
 	asm volatile("pushq %%rax\n\t"
 			"movl 0(%%rax),%%edx\n\t"
@@ -32,8 +43,8 @@ int SMIMonitor::i8k_smm(SMMRegisters *regs) {
 			"movl 4(%%rax),%%ebx\n\t"
 			"movl 8(%%rax),%%ecx\n\t"
 			"movl 12(%%rax),%%edx\n\t"
-			"movl 16(%%rax),%%esi\n\t"
-			"movl 20(%%rax),%%edi\n\t"
+ 			"movl 16(%%rax),%%esi\n\t"
+ 			"movl 20(%%rax),%%edi\n\t"
 			"popq %%rax\n\t"
 			"out %%al,$0xb2\n\t"
 			"out %%al,$0x84\n\t"
@@ -79,8 +90,8 @@ int SMIMonitor::i8k_smm(SMMRegisters *regs) {
 			: "%ebx", "%ecx", "%edx", "%esi", "%edi", "memory");
 #endif
 	
-	IOSimpleLockUnlock(preemptionLock);
-
+	atomic_store_explicit(&busy, false, memory_order_release);
+	
 	if ((rc != 0) || ((regs->eax & 0xffff) == 0xffff) || (regs->eax == eax)) {
 		return -1;
 	}
@@ -223,17 +234,35 @@ int SMIMonitor::i8k_set_fan(int fan, int speed) {
 }
 
 int SMIMonitor::i8k_set_fan_control_manual(int fan) {
+	// we have to write to both control registers since some Dell models
+	// support only one register and smm does not return error for unsupported one
 	SMMRegisters regs {};
 	regs.eax = I8K_SMM_IO_DISABLE_FAN_CTL1;
 	regs.ebx = (fan & 0xff);
-	return i8k_smm(&regs);
+	int result1 = i8k_smm(&regs);
+	
+	regs = {};
+	regs.eax = I8K_SMM_IO_DISABLE_FAN_CTL2;
+	regs.ebx = (fan & 0xff);
+	int result2 = i8k_smm(&regs);
+	
+	return (result1 >= 0) ? result1 : result2;
 }
 
 int SMIMonitor::i8k_set_fan_control_auto(int fan) {
+	// we have to write to both control registers since some Dell models
+	// support only one register and smm does not return error for unsupported one
 	SMMRegisters regs {};
 	regs.eax = I8K_SMM_IO_ENABLE_FAN_CTL1;
 	regs.ebx = (fan & 0xff);
-	return i8k_smm(&regs);
+	int result1 = i8k_smm(&regs);
+	
+	regs = {};
+	regs.eax = I8K_SMM_IO_ENABLE_FAN_CTL2;
+	regs.ebx = (fan & 0xff);
+	int result2 = i8k_smm(&regs);
+
+	return (result1 >= 0) ? result1 : result2;
 }
 
 void SMIMonitor::createShared() {
@@ -260,7 +289,6 @@ bool SMIMonitor::probe() {
 
 	bool success = true;
 
-	
 	while (!updateCall) {
 		updateCall = thread_call_allocate(staticUpdateThreadEntry, this);
 		if (!updateCall) {
@@ -268,7 +296,7 @@ bool SMIMonitor::probe() {
 			success = false;
 			break;
 		}
-		
+
 		IOLockLock(mainLock);
 		thread_call_enter(updateCall);
 		
@@ -293,7 +321,7 @@ bool SMIMonitor::probe() {
 			updateCall = nullptr;
 		}
 	}
-	
+
 	DBGLOG("sdell", "Based on I8kfan project and adopted to VirtualSMC plugin");
 
 	return success;
@@ -304,6 +332,14 @@ void SMIMonitor::start() {
 
 void SMIMonitor::handlePowerOff() {
 	if (awake) {
+		ignore_new_smc_updates = true;
+		if (fansStatus != 0) {
+			// turn off manual control for all fans
+			UInt16 data = 0;
+			postSmcUpdate(KeyFS__, -1, &data, sizeof(data), true);
+		}
+		DBGLOG("sdell", "SMIMonitor switched to sleep state, smc updates before sleep: %d", storedSmcUpdates.size());
+		while (storedSmcUpdates.size() != 0) { IOSleep(10); }
 		awake = false;
 	}
 }
@@ -311,16 +347,23 @@ void SMIMonitor::handlePowerOff() {
 void SMIMonitor::handlePowerOn() {
 	if (!awake) {
 		awake = true;
+		ignore_new_smc_updates = false;
+		DBGLOG("sdell", "SMIMonitor switched to awake state");
 	}
 }
 
-bool SMIMonitor::postSmcUpdate(SMC_KEY key, size_t index, const void *data, uint32_t dataSize)
+bool SMIMonitor::postSmcUpdate(SMC_KEY key, size_t index, const void *data, uint32_t dataSize, bool force_update)
 {
+	if (!force_update && (!awake || ignore_new_smc_updates)) {
+		DBGLOG("sdell", "SMIMonitor: postSmcUpdate for key %d has been ignored", key);
+		return false;
+	}
+	
 	IOSimpleLockLock(queueLock);
 
 	bool success = false;
 	while (1) {
-	
+
 		if (dataSize > sizeof(StoredSmcUpdate::data)) {
 			SYSLOG("sdell", "postRequest dataSize overflow %u", dataSize);
 			break;
@@ -363,20 +406,20 @@ bool SMIMonitor::postSmcUpdate(SMC_KEY key, size_t index, const void *data, uint
 
 IOReturn SMIMonitor::bindCurrentThreadToCpu0()
 {
-	// Obtain power management callbacks
+	// Obtain power management callbacks 10.7+
 	pmCallBacks_t callbacks {};
 	pmKextRegister(PM_DISPATCH_VERSION, nullptr, &callbacks);
-	
+
 	if (!callbacks.LCPUtoProcessor) {
 		SYSLOG("sdell", "failed to obtain LCPUtoProcessor");
 		return KERN_FAILURE;
 	}
-	
+
 	if (!callbacks.ThreadBind) {
 		SYSLOG("sdell", "failed to obtain ThreadBind");
 		return KERN_FAILURE;
 	}
-	
+
 	if (!IOSimpleLockTryLock(preemptionLock)) {
 		SYSLOG("sdell", "Preemption cannot be disabled before performing ThreadBind");
 		return KERN_FAILURE;
@@ -384,7 +427,7 @@ IOReturn SMIMonitor::bindCurrentThreadToCpu0()
 
 	bool success = true;
 	auto enable = ml_set_interrupts_enabled(FALSE);
-	
+
 	while (1)
 	{
 		auto processor = callbacks.LCPUtoProcessor(0);
@@ -513,12 +556,12 @@ void SMIMonitor::staticUpdateThreadEntry(thread_call_param_t param0, thread_call
 }
 
 void SMIMonitor::updateSensorsLoop() {
-	
+
 	for (int i=0; i<fanCount; ++i)
 		i8k_set_fan_control_auto(state.fanInfo[i].index); // force automatic control
-		
+
 	while (1) {
-		
+
 		for (int i=0; i<fanCount && awake; ++i)
 		{
 			int sensor = state.fanInfo[i].index;
@@ -541,7 +584,7 @@ void SMIMonitor::updateSensorsLoop() {
 			handleSmcUpdatesInIdle(4);
 		}
 		
-		handleSmcUpdatesInIdle(10);
+		handleSmcUpdatesInIdle(5);
 	}
 }
 
@@ -549,7 +592,7 @@ void SMIMonitor::handleSmcUpdatesInIdle(int idle_loop_count)
 {
 	for (int i=0; i<idle_loop_count; ++i)
 	{
-		if (awake) {
+		if (awake && storedSmcUpdates.size() != 0) {
 			IOSimpleLockLock(queueLock);
 			if (storedSmcUpdates.size() > 0) {
 				StoredSmcUpdate update = storedSmcUpdates[0];
@@ -573,14 +616,14 @@ void SMIMonitor::handleSmcUpdatesInIdle(int idle_loop_count)
 			}
 		}
 
-		IOSleep(50);
+		IOSleep(100);
 	}
 }
 
 void SMIMonitor::hanldeManualControlUpdate(size_t index, UInt8 *data)
 {
 	UInt16 val = data[0];
-	DBGLOG("sdell", "Set manual mode for fan %d to %d", index, val);
+	DBGLOG("sdell", "Set manual mode for fan %d to %s", index, val ? "enable" : "disable");
 
 	int rc = 0;
 	if (val != (fansStatus & (1 << index))>>index) {
@@ -589,7 +632,7 @@ void SMIMonitor::hanldeManualControlUpdate(size_t index, UInt8 *data)
 	}
 	if (rc == 0) {
 		fansStatus = val ? (fansStatus | (1 << index)) : (fansStatus & ~(1 << index));
-		DBGLOG("sdell", "Set manual mode for fan %d to %d, fansStatus = 0x%02x", index, val, fansStatus);
+		DBGLOG("sdell", "Set manual mode for fan %d to %s, global fansStatus = 0x%02x", index, val ? "enable" : "disable", fansStatus);
 	}
 	else
 		SYSLOG("sdell", "Set manual mode for fan %d to %d failed: %d", index, val, rc);
@@ -606,9 +649,21 @@ void SMIMonitor::hanldeManualTargetSpeedUpdate(size_t index, UInt8 *data)
 		int range = state.fanInfo[index].maxSpeed - state.fanInfo[index].minSpeed;
 		if (value > state.fanInfo[index].minSpeed + range/2)
 			status = 2;
-		int rc = i8k_set_fan(state.fanInfo[index].index, status);
-		if (rc != 0)
-			SYSLOG("sdell", "Set target speed for fan %d to %d failed: %d", index, value, rc);	}
+		else if (state.fanInfo[index].stopOffset != 0 && value < (state.fanInfo[index].minSpeed + state.fanInfo[index].stopOffset))
+			status = 0;		// stop fan
+		
+		int current_status = i8k_get_fan_status(state.fanInfo[index].index);
+		state.fanInfo[index].status = current_status;
+		if (current_status < 0 || current_status != status) {
+			int rc = i8k_set_fan(state.fanInfo[index].index, status);
+			if (rc != 0)
+				SYSLOG("sdell", "Set target speed for fan %d to %d failed: %d", index, value, rc);
+			else {
+				state.fanInfo[index].status = status;
+				DBGLOG("sdell", "Set target speed for fan %d to %d, status = %d", index, value, status);
+			}
+		}
+	}
 	else
 		SYSLOG("sdell", "Set target speed for fan %d to %d ignored since auto control is active", index, value);
 }